NaN problem is fixed
In the previous version, the boundary condition could evaluate to NaN if
the simulation time Ny is sufficiently long. The reason is due to the
numerical instability of how the incomplete gamma function is implemented.

In this version, this problem is completely solved by implementing a stabler
function following Numerical Recipes. Also, the usage of tgamma is eliminated
to the max extent in order to prevent overflow. Hopefully this can help
access a wider range on the 1+1D plane.