the program now accepts options (save_chi and save_psi)
If the code cannot find the value for a given key, a null pointer is
returned. If this happens to mandatory parameters, a seg fault will
happen (we need this!). If this happens to opions (see below), then
the code shall handle it by taking the default value.

Currently two options can be given:
  1. save_chi: save the result of the two-photon wavefunction
  2. save_psi: save the solution of the delay PDE
Both default to off (=0) to prevent from eating up the disk space,
but at least one of them should be set to on (=1) in order not to
run the code for nothing; this check is added to sanity_check.

The script generate_input.py and the example input k0a_0.5pi_on_new
are updated accordingly.