barrier removed --- STILL BUGGY, SEE NOTE BELOW!!!
Now solvers are not "batched" by the barrier, and a cylic relation is
hold among them; however, on my laptop occasionally the current
implementation would produce incorrect result. To reproduce this
behavior, follow the procedure below:

1. make clean
2. make -j 4
3. run the simulation with multiple threads (Nth>1)

After this incorrect run, it seems the future runs will be correct.
Apparently synchronization is somehow buggy and is under investigation.