set Nth as the number of threads for both pthreads and openmp
OpenMP's environmental variable OMP_NUM_THREADS will be neglected if it
conflicts with Nth specified in the input.

TODO: think about the "thread-doubling" issue when supports for pthreads
and openmp co-exist. Most likely, we need to create the OpenMP team in
another thread, and then kill that thread to kill the team before the
pthreads solvers begin to work. (But then after that we need to recreate
the team...)